Move `WolfSession` into its own `session.rs`
